+++ src/views/ConfigEditor/FormParts/DefaultValues/DefaultValues.tsx
@@ -13,12 +13,7 @@ interface DefaultValuesInterface {
   onFieldChange: any;
 }
 
-export const DefaultValues = ({
-  jsonData,
-  newOptions,
-  onSwitchToggle,
-  onFieldChange,
-}: DefaultValuesInterface) => {
+export const DefaultValues = ({ jsonData, newOptions, onSwitchToggle, onFieldChange }: DefaultValuesInterface) => {
   const [defaultDateTime64Options, setDefaultDateTime64Options] = useState<any[]>([]);
   const [defaultDateTimeOptions, setDefaultDateTimeOptions] = useState<any[]>([]);
   const [defaultUint32Options, setDefaultUint32Options] = useState<any[]>([]);
@@ -185,6 +180,9 @@ export const DefaultValues = ({
                 { label: 'DateTime64', value: TimestampFormat.DateTime64 },
                 { label: 'TimeStamp', value: TimestampFormat.TimeStamp },
                 { label: 'Float', value: TimestampFormat.Float },
+                { label: 'Timestamp64(3)', value: TimestampFormat.TimeStamp64_3 },
+                { label: 'Timestamp64(6)', value: TimestampFormat.TimeStamp64_6 },
+                { label: 'Timestamp64(9)', value: TimestampFormat.TimeStamp64_9 },
               ]}
               value={jsonData.defaultDateTimeType}
             />
@@ -241,6 +239,18 @@ export const DefaultValues = ({
                 onFieldChange({ value: changeEvent ? changeEvent.value : undefined }, 'defaultDateDate32');
               }}
               options={defaultDateDate32Options}
+            />
+          </InlineField>
+          <h6>Logs settings</h6>
+          <InlineField label="Context window" labelWidth={32} style={{ marginLeft: '30px' }}>
+            <Select
+              width={24}
+              data-testid="context-window-size-select"
+              onChange={(changeEvent) => {
+                onFieldChange({ value: changeEvent ? changeEvent.value : undefined }, 'contextWindowSize');
+              }}
+              options={['10', '20', '50', '100'].map((value) => ({ label: value + ' entries', value }))}
+              value={jsonData.contextWindowSize}
             />
           </InlineField>
         </>