+++ src/views/QueryEditor/components/QueryBuilder/hooks/useConnectionData.ts
@@ -74,6 +74,21 @@ export const useConnectionData = (query, datasource) => {
             "type = 'UInt32' " +
             'ORDER BY name';
           break;
+        case TimestampFormat.TimeStamp64_3:
+        case TimestampFormat.TimeStamp64_6:
+        case TimestampFormat.TimeStamp64_9:
+          query =
+            'SELECT name ' +
+            'FROM system.columns ' +
+            "WHERE database = '" +
+            selectedDatabase +
+            "' AND " +
+            "table = '" +
+            selectedTable +
+            "' AND " +
+            "(type LIKE 'UInt%' OR type LIKE 'Int%')" +
+            'ORDER BY name';
+          break;
         case TimestampFormat.Float:
           query =
             'SELECT name ' +