+++ tests/testflows/steps/connections/datasources/altinity_edit/locators.py
@@ -190,6 +190,23 @@ class Locators:
         return driver.find_element(SelectBy.XPATH,
                                    f'//*[./text()="Date Field"]/..//input[contains(@id, "react-select")]')
 
+    @property
+    def context_window_field(self):
+        driver: WebDriver = current().context.driver
+        return driver.find_element(SelectBy.XPATH,
+                                   f'//*[./text()="Context window"]/..//input[contains(@id, "react-select")]')
 
+
+    @property
+    def configure_adhoc_filter_request(self):
+        driver: WebDriver = current().context.driver
+        return driver.find_element(SelectBy.XPATH,
+                                   f'//*[@data-testid="data-testid Code editor container"]//*[@class="view-lines monaco-mouse-cursor-text"]')
+
+    @property
+    def configure_adhoc_filter_request_input(self):
+        driver: WebDriver = current().context.driver
+        return driver.find_element(SelectBy.XPATH,
+                                   f'//*[@data-testid="data-testid Code editor container"]//textarea')
 
 locators = Locators()